Further improved on placement of towers
Made it so the tower will snap to whatever grid the mouse is inside of.
I did this by making every grid into a rect and looping over all grids and using collidepoint() to see if the mouse collided with a rect.
Then the tower would be moved to that rect.